Add files via upload
In this project, we analyze flight data to gain insights into the performance and reliability of airlines. Initially, we obtained data via an API. We created SQL tables to structure the data and loaded it for preliminary processing and aggregation, focusing on metrics such as flight delays and reliability by route, year, and day.

We also utilized Python, specifically the pandas library, to query the data and prepare it for visualization. Using Plotly, Seaborn, and Dash, we created interactive dashboards to present our findings. The dashboards feature a table of the top 10 most reliable routes, bar charts comparing reliability year-over-year, and line graphs illustrating daily reliability trends.

Navigation links within the Dash application allow users to switch between different reports. This project effectively integrates SQL for data processing with Python and Dash for data visualization, providing a comprehensive tool for analyzing and understanding flight performance over time.